- Moved existing msg file reading functions and structs from HeroAppearance.cpp to a dedicated Message.h/.cpp file pair. - Added extern map to hold read extra msg files.
